Updated version to latest release
Updated the context and intents example now that view names specified in snapshots are respected and persist. This removes the need for inspecting the DOM to find the assigned names.
Updated the web client example to re-use more code and to check to see if the document is ready as core-web needs to assign an iframe to the document. The goal of the example was to show a way of assigning a script tag or use import across content that might belong to multiple hosts and supports container and web. In order to not rely on the developer using defer or other approaches the example now uses checks to see if the document is ready.