Added an example of a context group picker
Added the context group picker that will be on frontend-framework-starter (using stencil) and used it from a platform view and a content view. Added some additional logic so that the component doesn't need to worry about importing the fin or fdc3 api it just uses them. Also added some additional options to the context group picker to support some scenarios such as a mix use (platform and content).